requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langsmith
openai
mysql-connector-python
python-dotenv
sqlalchemy